+++ app/src/main/java/com/example/watchlist2/data/remote/dto/AnimeSearchResultsDto.kt
@@ -20,7 +20,7 @@ data class AnimeSearchResultsDto(
     @JsonClass(generateAdapter = true)
     data class AnimeSearchResultDto(
         @Json(name = "mal_id")
-        val malId: Int,
+        val malId: Long,
         @Json(name = "url")
         val url: String?,
         @Json(name = "image_url")
